# tests/golden/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoightKampff OCR API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/verification/passport/ocr": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Passport Ocr",
        "operationId": "passport_ocr_verification_passport_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_passport_ocr_verification_passport_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/id-card/ocr": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Id-Card Ocr",
        "operationId": "id_card_ocr_verification_id_card_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_id_card_ocr_verification_id_card_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationIdCardOcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/driving-licence/ocr": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Driving-Licence Ocr",
        "operationId": "driving_licence_ocr_verification_driving_licence_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_driving_licence_ocr_verification_driving_licence_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/passport/ocr/batch": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Ordinary Batch Route",
        "operationId": "passport_verification_ocr_batch",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_passport_verification_ocr_batch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/id-card/ocr/batch": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Id Batch",
        "operationId": "id_card_verification_ocr_batch",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_id_card_verification_ocr_batch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/driving-licence/ocr/batch": {
      "post": {
        "tags": [
          "Verification OCR"
        ],
        "summary": "Driving Batch Route",
        "operationId": "driving_licence_verification_ocr_batch",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_driving_licence_verification_ocr_batch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOcrBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/passport/check": {
      "post": {
        "tags": [
          "Verification checks"
        ],
        "summary": "Passport Check",
        "operationId": "passport_check_verification_passport_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/driving-licence/check": {
      "post": {
        "tags": [
          "Verification checks"
        ],
        "summary": "Driving Check",
        "operationId": "driving_check_verification_driving_licence_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verification/id-card/check": {
      "post": {
        "tags": [
          "Verification checks"
        ],
        "summary": "Id Check",
        "operationId": "id_check_verification_id_card_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationIdCardCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_driving_licence_ocr_verification_driving_licence_ocr_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_driving_licence_ocr_verification_driving_licence_ocr_post"
      },
      "Body_driving_licence_verification_ocr_batch": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "format": "binary",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "Select one or more image files, ZIP archives, or a mixture of both. ZIP archives may contain images in nested folders.",
            "default": []
          },
          "archive": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive"
          }
        },
        "type": "object",
        "title": "Body_driving_licence_verification_ocr_batch"
      },
...
      "Body_id_card_ocr_verification_id_card_ocr_post": {
        "properties": {
          "front": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "front",
          "back"
        ],
        "title": "Body_id_card_ocr_verification_id_card_ocr_post"
      },
...
      "Body_id_card_verification_ocr_batch": {
        "properties": {
          "archive": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archive"
          }
        },
        "type": "object",
        "required": [
          "archive"
        ],
        "title": "Body_id_card_verification_ocr_batch"
      },
...
      "Body_passport_ocr_verification_passport_ocr_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_passport_ocr_verification_passport_ocr_post"
      },
...
      "Body_passport_verification_ocr_batch": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "format": "binary",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "Select one or more image files, ZIP archives, or a mixture of both. ZIP archives may contain images in nested folders.",
            "default": []
          },
          "archive": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive"
          }
        },
        "type": "object",
        "title": "Body_passport_verification_ocr_batch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VerificationCheckRequest": {
        "properties": {
          "ocr": {
            "$ref": "#/components/schemas/VerificationOcrResponse"
          },
          "fields": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "type": "object",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "ocr",
          "fields"
        ],
        "title": "VerificationCheckRequest"
      },
...
      "VerificationIdCardOcrResponse": {
        "properties": {
          "front": {
            "items": {
              "$ref": "#/components/schemas/VerificationOcrLine"
            },
            "type": "array",
            "title": "Front"
          },
          "back": {
            "items": {
              "$ref": "#/components/schemas/VerificationOcrLine"
            },
            "type": "array",
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "front",
          "back"
        ],
        "title": "VerificationIdCardOcrResponse"
      },
...
      "VerificationOcrBatchResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "succeeded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Succeeded"
          },
          "failed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/VerificationOcrBatchItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "succeeded",
          "failed",
          "items"
        ],
        "title": "VerificationOcrBatchResponse"
      },
...
      "VerificationOcrResponse": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/VerificationOcrLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "VerificationOcrResponse"
      },
      "VerificationResponse": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/VerificationFieldResult"
            },
            "type": "object",
            "title": "Fields"
          },
          "summary": {
            "$ref": "#/components/schemas/VerificationSummary"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "summary"
        ],
        "title": "VerificationResponse"
      },
...
    }
  },
...
}